Code refactored Single Responsibility Principle Addressed.
