+++ toll_fee_calc_app/src/main/java/com/evolve_technology/calculator/service/TollFeeService.java
@@ -1,0 +1,7 @@
+package com.evolve_technology.calculator.service;
+
+public interface TollFeeService {
+
+	public int getTollFee(final String date, String vehicle);
+
+}
