+++ toll_fee_calc_app/src/main/java/com/evolve_technology/calculator/service/TollFreeDatesService.java
@@ -1,0 +1,8 @@
+package com.evolve_technology.calculator.service;
+
+import java.util.List;
+
+public interface TollFreeDatesService {
+	public Boolean isTollFreeDate(String date);
+	public List<String> getTollFreeDates();
+}
