+++ toll_fee_calc_app/src/main/java/com/evolve_technology/calculator/service/TollFreeVehicleService.java
@@ -1,0 +1,8 @@
+package com.evolve_technology.calculator.service;
+
+import java.util.List;
+
+public interface TollFreeVehicleService {
+	public List<String> getTollFreeVehicles();
+	public boolean isTollFreeVehicle(String vehicle);
+}
